refactor arguments_converter (split subpackage)
