+++ lib/freecad_stub_gen/generators/common/doc_string.py
@@ -8,7 +8,7 @@ from xml.etree import ElementTree as ET
 
 from freecad_stub_gen.generators.common.annotation_parameter import AnnotationParam, RawRepr, \
     SelfSignature
-from freecad_stub_gen.generators.common.arguments_converter import DEFAULT_ARG_NAME
+from freecad_stub_gen.generators.common.arguments_converter.definitions import DEFAULT_ARG_NAME
 from freecad_stub_gen.generators.common.cpp_function import findFunctionCall, \
     generateExpressionUntilChar
 from freecad_stub_gen.generators.common.names import validatePythonValue
